+++ dbos/dbos.py
@@ -28,6 +28,7 @@ from opentelemetry.trace import Span
 from dbos.core import (
     _execute_workflow_uuid,
     _start_workflow,
+    _transaction,
     _workflow_wrapper,
     _WorkflowHandlePolling,
 )
@@ -217,120 +218,7 @@ class DBOS:
     def transaction(
         self, isolation_level: IsolationLevel = "SERIALIZABLE"
     ) -> Callable[[F], F]:
-        def decorator(func: F) -> F:
-            def invoke_tx(*args: Any, **kwargs: Any) -> Any:
-                with self.app_db.sessionmaker() as session:
-                    attributes: TracedAttributes = {
-                        "name": func.__name__,
-                        "operationType": OperationType.TRANSACTION.value,
-                    }
-                    with EnterDBOSTransaction(session, attributes=attributes) as ctx:
-                        txn_output: TransactionResultInternal = {
-                            "workflow_uuid": ctx.workflow_uuid,
-                            "function_id": ctx.function_id,
-                            "output": None,
-                            "error": None,
-                            "txn_snapshot": "",  # TODO: add actual snapshot
-                            "executor_id": None,
-                            "txn_id": None,
-                        }
-                        retry_wait_seconds = 0.001
-                        backoff_factor = 1.5
-                        max_retry_wait_seconds = 2.0
-                        while True:
-                            has_recorded_error = False
-                            try:
-                                with session.begin():
-                                    # This must be the first statement in the transaction!
-                                    session.connection(
-                                        execution_options={
-                                            "isolation_level": isolation_level
-                                        }
-                                    )
-                                    # Check recorded output for OAOO
-                                    recorded_output = (
-                                        ApplicationDatabase.check_transaction_execution(
-                                            session,
-                                            ctx.workflow_uuid,
-                                            ctx.function_id,
-                                        )
-                                    )
-                                    if recorded_output:
-                                        if recorded_output["error"]:
-                                            deserialized_error = utils.deserialize(
-                                                recorded_output["error"]
-                                            )
-                                            has_recorded_error = True
-                                            raise deserialized_error
-                                        elif recorded_output["output"]:
-                                            return utils.deserialize(
-                                                recorded_output["output"]
-                                            )
-                                        else:
-                                            raise Exception(
-                                                "Output and error are both None"
-                                            )
-                                    output = func(*args, **kwargs)
-                                    txn_output["output"] = utils.serialize(output)
-                                    assert (
-                                        ctx.sql_session is not None
-                                    ), "Cannot find a database connection"
-                                    ApplicationDatabase.record_transaction_output(
-                                        ctx.sql_session, txn_output
-                                    )
-                                    break
-                            except DBAPIError as dbapi_error:
-                                if dbapi_error.orig.pgcode == "40001":  # type: ignore
-                                    # Retry on serialization failure
-                                    DBOS.span.add_event(
-                                        "Transaction Serialization Failure",
-                                        {"retry_wait_seconds": retry_wait_seconds},
-                                    )
-                                    time.sleep(retry_wait_seconds)
-                                    retry_wait_seconds = min(
-                                        retry_wait_seconds * backoff_factor,
-                                        max_retry_wait_seconds,
-                                    )
-                                    continue
-                                raise dbapi_error
-                            except Exception as error:
-                                # Don't record the error if it was already recorded
-                                if not has_recorded_error:
-                                    txn_output["error"] = utils.serialize(error)
-                                    self.app_db.record_transaction_error(txn_output)
-                                raise error
-                return output
-
-            fi = get_or_create_func_info(func)
-
-            @wraps(func)
-            def wrapper(*args: Any, **kwargs: Any) -> Any:
-                rr: Optional[str] = check_required_roles(func, fi)
-                # Entering transaction is allowed:
-                #  In a workflow (that is not in a transaction / comm already)
-                #  Not in a workflow (we will start the single op workflow)
-                ctx = get_local_dbos_context()
-                if ctx and ctx.is_within_workflow():
-                    assert (
-                        ctx.is_workflow()
-                    ), "Transactions must be called from within workflows"
-                    with DBOSAssumeRole(rr):
-                        return invoke_tx(*args, **kwargs)
-                else:
-                    tempwf = self.workflow_info_map.get("<temp>." + func.__qualname__)
-                    assert tempwf
-                    return tempwf(*args, **kwargs)
-
-            def temp_wf(*args: Any, **kwargs: Any) -> Any:
-                return wrapper(*args, **kwargs)
-
-            wrapped_wf = _workflow_wrapper(self, temp_wf)
-            set_dbos_func_name(temp_wf, "<temp>." + func.__qualname__)
-            self._register_wf_function(get_dbos_func_name(temp_wf), wrapped_wf)
-
-            return cast(F, wrapper)
-
-        return decorator
+        return _transaction(self, isolation_level)
 
     # Mirror the CommunicatorConfig from TS. However, we disable retries by default.
     def communicator(
